Tweaks to compile with Visual Studio

+++ CarveLookAndFeel.cpp
@@ -29,11 +29,11 @@ void CarveLookAndFeel::drawRotarySlider(Graphics& g,
                                         float rotaryEndAngle,
                                         Slider &slider) {
     // calculate useful constants
-    const float rangeOfMotion {260 * (M_PI / 180)};
+    const double rangeOfMotion {260 * (M_PI / 180)};
     const double rotation {((slider.getValue() - slider.getMinimum()) / (slider.getMaximum() - slider.getMinimum())) * rangeOfMotion - rangeOfMotion / 2};
 
     const int margin {1};
-    const float diameter {static_cast<float>(height - margin * 2)};
+    const double diameter {static_cast<double>(height - margin * 2)};
     
     // draw centre circle
     Path p;
@@ -50,7 +50,7 @@ void CarveLookAndFeel::drawRotarySlider(Graphics& g,
     
     p.clear();
     
-    const float gap {0.4};
+    const double gap {0.4};
     p.addCentredArc(width / 2, height / 2, diameter / 2, diameter / 2, rotation, gap, 2 * M_PI - gap, true);
     
     g.strokePath(p, PathStrokeType(2.0f));
